Move to conventional project structure with pages
